skills/Auto-Redbook-Skills/requirements.txt
markdown>=3.4.0
PyYAML>=6.0
playwright>=1.40.0
python-dotenv>=1.0.0
httpx>=0.22.0
xhs
